refactor: move relevant code to Core namespace
